requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
torch
sentence-transformers
fastapi
uvicorn
rank_bm25
openai
anthropic
# ragas  # Phase 3 — requires llvmlite/numba, install later
structlog
pyyaml
pypdf
unstructured
langfuse
